factor out variables into own file to declutter files
