--- inlang/packages/paraglide/paraglide-js/src/compiler/runtime/base-locale.js
@@ -1,9 +1,0 @@
-/**
- * The project's base locale.
- *
- * @example
- *   if (locale === baseLocale) {
- *     // do something
- *   }
- */
-export const baseLocale = "<base-locale>";
