--- inlang/packages/paraglide/paraglide-js/src/compiler/runtime/cookie-name.js
@@ -1,2 +1,0 @@
-/** @type {string} */
-export const cookieName = "<cookie-name>";
