--- inlang/packages/paraglide/paraglide-js/src/compiler/runtime/locales.js
@@ -1,9 +1,0 @@
-/**
- * The project's locales that have been specified in the settings.
- *
- * @example
- *   if (locales.includes(userSelectedLocale) === false) {
- *     throw new Error('Locale is not available');
- *   }
- */
-export const locales = /** @type {const} */ (["<base-locale>"]);
