--- inlang/packages/paraglide/paraglide-js/src/compiler/runtime/strategy.js
@@ -1,4 +1,0 @@
-/**
- * @type {Array<"cookie" | "baseLocale" | "pathname" | "variable">}
- */
-export const strategy = ["variable"];
